notification-alert/requirements.txt
fastapi==0.128.0
fastapi-cli==0.0.20
fastapi-cloud-cli==0.9.0
PyJWT==2.8.0
pydantic==2.12.5
pydantic-extra-types==2.11.0
pydantic-settings==2.12.0
pydantic_core==2.41.5
redis==7.1.0
uvicorn==0.40.0